Linear search

+++ 0x1E-search_algorithms/0-linear.c
@@ -24,7 +24,7 @@ int linear_search(int *array, size_t size, int value)
 		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 		{
-			return (i);
+			return ((int)i);
 		}
 	}
 	return (-1);
